@@ -16,9 +16,10 @@
     "cy:run-ct": "cypress run --component"
   },
   "dependencies": {
+    "@popperjs/core": "^2.11.6",
     "axios": "^1.2.0",
-    "bootstrap": "^4.6.0",
-    "bootstrap-vue": "^2.23.1",
+    "bootstrap": "^5.2.2",
+    "bootstrap-vue-3": "^0.3.12",
     "bootswatch": "^5.2.2",
     "core-js": "^3.26.1",
     "date-fns": "^2.29.3",
@@ -28,14 +29,15 @@
     "humanize-duration": "^3.27.3",
     "pinia": "^2.0.27",
     "pinia-plugin-persistedstate": "^3.0.1",
-    "vue": "^2.7.14",
+    "vue": "3.2.42",
     "vue-class-component": "^7.2.5",
     "vue-demi": "0.13.11",
     "vue-echarts": "^6.2.3",
     "vue-grid-layout": "^2.4.0",
     "vue-material-design-icons": "^5.1.2",
-    "vue-router": "^3.6.5",
-    "vue-select": "^3.20.0"
+    "vue-router": "^4.1.6",
+    "vue-select": "^4.0.0-beta.5",
+    "vue3-drr-grid-layout": "^1.9.6"
   },
   "devDependencies": {
     "@cypress/vite-dev-server": "^4.0.1",
@@ -44,8 +46,8 @@
     "@types/jest": "^27.5.0",
     "@typescript-eslint/eslint-plugin": "^2.33.0",
     "@typescript-eslint/parser": "^5.45.0",
-    "@vitejs/plugin-vue2": "^1.1.2",
-    "@vue/composition-api": "^1.7.1",
+    "@vitejs/plugin-vue": "^1.2.2",
+    "@vue/compiler-sfc": "3.2.42",
     "@vue/eslint-config-prettier": "^6.0.0",
     "@vue/eslint-config-typescript": "^5.1.0",
     "@vue/runtime-dom": "^3.2.45",
@@ -56,7 +58,6 @@
     "eslint-plugin-vue": "^7.20.0",
     "jest": "^27.5.1",
     "mutationobserver-shim": "^0.3.7",
-    "popper.js": "^1.16.1",
     "portal-vue": "^2.1.7",
     "prettier": "^2.8.0",
     "sass": "^1.56.1",
@@ -65,7 +66,6 @@
     "typescript": "~4.9.3",
     "vite": "^2.9.14",
     "vite-jest": "^0.1.4",
-    "vue-template-compiler": "^2.7.14",
     "vue-tsc": "^1.0.10"
   }
 }
