@@ -1,7 +1,7 @@
-import Vue from 'vue';
-
-import BootstrapVue from 'bootstrap-vue';
+import BootstrapVue3 from 'bootstrap-vue-3';
+import 'bootstrap/dist/css/bootstrap.css';
+import 'bootstrap-vue-3/dist/bootstrap-vue-3.css';
 
 import '@/styles/main.scss';
 
-Vue.use(BootstrapVue);
+export { BootstrapVue3 };
